[NO-JIRA] fix: let guests see the daily fortune without logging in
배포하고 브라우저로 직접 눌러보니 메인 CTA 가 곧바로 로그인 벽으로 갔다.
방문자가 볼 수 있는 게 0개였고, 매직링크는 Supabase Redirect URL 허용목록에
웹 주소가 없어 돌아오지도 않는다. HTTP 200 만 보고 "동작한다"고 판단한 것이
잘못이었다 — 로그인 벽도 200 을 낸다.

- 운세 화면을 인증 게이트 밖으로 꺼냈다: src/features/daily/ 로 옮기고
  공개 라우트 /운세/오늘 을 만들었다. 한글 slug 는 계획서 §5.2 결정을 따른다.
- /app/f/daily 는 리다이렉트만 남긴다.
- 랜딩 CTA 를 공개 경로로 돌리고 sitemap 에 추가했다 (한글 경로는 encodeURI).
- 제출 시 세션이 없으면 signInAnonymously 를 시도하고, 프로젝트에서 익명
  로그인이 꺼져 있으면 anon key 로 그대로 진행한다. 서버가 401 을 주면 기존
  auth 분기가 로그인 안내를 띄운다.

결과 앞에 벽을 두지 않는 건 계획서 §8.1 의 퍼널 설계이자,
resolveChatOnboardingGate 가 비로그인에 'ready' 를 주는 네이티브 정책
(Apple 5.1.1(v)) 과도 일관된다.

브라우저 실검증: 랜딩 -> "오늘의 운세 보기" -> 생년월일 입력 -> 제출 ->
92점 결과가 6개 항목 + 조언/주의/행운요소/할일 까지 전부 렌더됨.
